requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-mcp-adapters>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
